finish politicas
