--- src/components/ItemListParagraph/ItemListParagraph.jsx
@@ -1,8 +1,0 @@
-import styles from "./styles.module.css";
-
-const ItemListParagraph = (props) => {
-  console.log(props.list);
-  return <li>{props.list}</li>;
-};
-
-export default ItemListParagraph;
