--- src/components/ItemParagraph/ItemParagraph.jsx
@@ -1,7 +1,0 @@
-import styles from "./style.module.css";
-
-const ItemParagraph = (props) => {
-  return <p className={styles.ItemParagraph}>{props.paragraph}</p>;
-};
-
-export default ItemParagraph;
