+++ src/pages/politicas/styles.module.css
@@ -1,32 +1,32 @@
-.PolitContainer {
+.politContainer {
   margin-left: 30px;
 }
 
-.PolitContainer h1 {
+.politContainer h1 {
   margin-bottom: 5px;
 }
 
-.IntroductionContainer {
+.introductionContainer {
   font-size: 10px;
 }
 
 @media (min-width: 576px) {
-  .PolitContainer h1 {
+  .politContainer h1 {
     margin-bottom: 5px;
   }
 }
 
 @media (min-width: 768px) {
-  .PolitContainer {
+  .politContainer {
     margin-left: 30px;
   }
 
-  .PolitContainer h1 {
+  .politContainer h1 {
     margin-top: 20px;
     margin-bottom: 50px;
   }
 
-  .IntroductionContainer {
+  .introductionContainer {
     font-size: 20px;
   }
 }
